-- txn model

+++ backend/src/models/transaction.rs
@@ -1,4 +1,5 @@
 
+use bigdecimal::BigDecimal;
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 use sqlx::prelude::FromRow;
@@ -10,7 +11,7 @@ pub struct Transaction {
     pub user_id:            Uuid,
     pub device_id:          Uuid,
     pub amount_kobo:        i64,
-    pub units_purchased:    f64,
+    pub units_purchased:    BigDecimal,
     pub channel:            String,
     pub status:             String,
     pub interswitch_ref:    Option<String>,
